requirements.txt
pytest==8.0.2
requests==2.31.0
selene==2.0.0rc9
allure-pytest==2.13.5
curlify==2.2.1